c/a/tests: Extract async_data_uploader_fixture
