+++ src/v/cluster/archival/tests/async_data_uploader_fixture.cc
@@ -1,0 +1,18 @@
+/*
+ * Copyright 2025 Redpanda Data, Inc.
+ *
+ * Licensed as a Redpanda Enterprise file under the Redpanda Community
+ * License (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * https://github.com/redpanda-data/redpanda/blob/master/licenses/rcl.md
+ *
+ */
+
+#include "cluster/archival/tests/async_data_uploader_fixture.h"
+
+namespace archival {
+
+ss::logger adu_fixture_log("async-uploader-fixture");
+
+} // namespace archival
